analysis/requirements.txt
bibtexparser==1.2.0
dictdiffer==0.8.1
fuzzywuzzy==0.18.0
GitPython==3.1.14
langdetect==1.0.8
lxml==4.0
pandas==1.1.5
pdfminer.six==20201018
pre-commit==2.12.1
pytest==6.2.4
pytest-cov==2.12.0
python-Levenshtein==0.12.2
requests==2.25.1
tqdm==4.55.0
